+++ enterprise/cmd/executor/internal/config/config_test.go
@@ -70,6 +70,8 @@ func TestConfig_Load(t *testing.T) {
 			return "sourcegraph/step-image:latest"
 		case "KUBERNETES_JOB_ANNOTATIONS":
 			return `{"foo": "bar", "faz": "baz"}`
+		case "KUBERNETES_JOB_POD_ANNOTATIONS":
+			return `{"foo": "bar", "faz": "baz"}`
 		case "KUBERNETES_IMAGE_PULL_SECRETS":
 			return "foo,bar"
 		default:
@@ -172,9 +174,15 @@ func TestConfig_Load(t *testing.T) {
 		cfg.KubernetesAdditionalJobVolumeMounts,
 	)
 	assert.Equal(t, "sourcegraph/step-image:latest", cfg.KubernetesSingleJobStepImage)
+
 	assert.Len(t, cfg.KubernetesJobAnnotations, 2)
 	assert.Equal(t, "bar", cfg.KubernetesJobAnnotations["foo"])
 	assert.Equal(t, "baz", cfg.KubernetesJobAnnotations["faz"])
+
+	assert.Len(t, cfg.KubernetesJobPodAnnotations, 2)
+	assert.Equal(t, "bar", cfg.KubernetesJobPodAnnotations["foo"])
+	assert.Equal(t, "baz", cfg.KubernetesJobPodAnnotations["faz"])
+
 	assert.Equal(t, "foo,bar", cfg.KubernetesImagePullSecrets)
 }
 
@@ -229,6 +237,7 @@ func TestConfig_Load_Defaults(t *testing.T) {
 	assert.Empty(t, cfg.KubernetesAdditionalJobVolumeMounts)
 	assert.Equal(t, "sourcegraph/batcheshelper:insiders", cfg.KubernetesSingleJobStepImage)
 	assert.Nil(t, cfg.KubernetesJobAnnotations)
+	assert.Nil(t, cfg.KubernetesJobPodAnnotations)
 	assert.Empty(t, cfg.KubernetesImagePullSecrets)
 }